# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"Gurukul AI Backend","description":"Secure AI Orchestrator and Chapter Processing Engine","version":"1.0.0"},"paths":{"/api/health/":{"get":{"tags":["Health"],"summary":"Health Check","operationId":"health_check_api_health__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/health/ai":{"get":{"tags":["Health"],"summary":"Ai Health","operationId":"ai_health_api_health_ai_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/ai/generate":{"post":{"tags":["AI"],"summary":"Generate","operationId":"generate_api_ai_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/ai/generate-structured":{"post":{"tags":["AI"],"summary":"Generate Structured","operationId":"generate_structured_api_ai_generate_structured_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StructuredGenerateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/ai/embeddings":{"post":{"tags":["AI"],"summary":"Generate Embeddings","operationId":"generate_embeddings_api_ai_embeddings_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmbeddingRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/chapters/process":{"post":{"tags":["Chapters"],"summary":"Start Chapter Processing","description":"Accept exactly ONE PDF for ONE chapter.\n\nIf the student already has an active chapter job, reject the request\nbefore permanently keeping the newly uploaded file.","operationId":"start_chapter_processing_api_chapters_process_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_start_chapter_processing_api_chapters_process_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/chapters/status/{job_id}":{"get":{"tags":["Chapters"],"summary":"Get Job Status","operationId":"get_job_status_api_chapters_status__job_id__get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_start_chapter_processing_api_chapters_process_post":{"properties":{"student_id":{"type":"string","title":"Student Id"},"book_id":{"type":"string","title":"Book Id"},"chapter_id":{"type":"string","title":"Chapter Id"},"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["student_id","book_id","chapter_id","file"],"title":"Body_start_chapter_processing_api_chapters_process_post"},"EmbeddingRequest":{"properties":{"text":{"type":"string","title":"Text"}},"type":"object","required":["text"],"title":"EmbeddingRequest"},"GenerateRequest":{"properties":{"prompt":{"type":"string","title":"Prompt"},"task_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Task Type","default":"general"}},"type":"object","required":["prompt"],"title":"GenerateRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"StructuredGenerateRequest":{"properties":{"prompt":{"type":"string","title":"Prompt"},"schema":{"additionalProperties":true,"type":"object","title":"Schema"},"task_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Task Type","default":"structured"}},"type":"object","required":["prompt","schema"],"title":"StructuredGenerateRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}